Complete refactoring and separating code concerns into different JS modules
